Created the list of characters
I searched the web for a list of characters and the best that I found was a table. I used the get data from url function from excel to get the data and then used vs code's multi cursor to make it all into one string. I then made a JavaScript algorithm to make it an array.