First 10 characters finished!!
If you are lucky enough to get a match between two of the first 10 characters on the list then you will have 2 beautiful images to go along the names. I also edited the list and will continue to take out the names of characters I don't want to appear on the website. I originally had a list of ALL characters but will boil it down to characters that will mean something to the user or will at least be recognized. (By me at least)